third_party/blink/web_tests/platform/generic/fast/forms/text/input-maxlength-ime-preedit-expected.txt

This test checks users can input pre-edit text longer than maxlength.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getSelection().toString() is "abcd"
PASS successfullyParsed is true

TEST COMPLETE
